added RBAC test and did house keeping according to PEP 8 style guide

+++ test_app.py
@@ -48,42 +48,62 @@ class CapstoneTestCase(unittest.TestCase):
     
     """
     Tests to be executed
-    """
-        
-    #Test POST /movies
-    def test_create_new_movie(self):
+    """  
+    
+    #Test POST /movies --RBAC - Error (Casting Assistant is not allowed to perform POST /movies)===========================
+    def test_create_new_movie_RBAC_error(self):
+        token_assistant = os.environ['TOKEN_ASSISTANT']
+        res = self.client().post('/movies', json={"title": "Men in Black2", "date": "2002"}, headers={"Authorization":"Bearer "+token_assistant})
+        data = json.loads(res.data.decode('utf-8'))
+        self.assertEqual(res.status_code, 403)
+    
+
+    #Test POST /movies --Authorization - Error ===========================
+    def test_create_new_movie_authorization_error(self):
         res = self.client().post('/movies', json={"title": "Men in Black2", "date": "2002"})
+        data = json.loads(res.data.decode('utf-8'))
+        self.assertEqual(res.status_code, 401)
+
+
+    #Test POST /movies ===========================
+    def test_create_new_movie(self):
+        token_Producer = os.environ['TOKEN_PRODUCER']
+        res = self.client().post('/movies', json={"title": "Men in Black2", "date": "2002"}, headers={"Authorization":"Bearer "+token_Producer})
         data = json.loads(res.data.decode('utf-8'))
         self.assertEqual(res.status_code, 201)
         self.assertEqual(data['success'], True)
 
     #Test POST /movies - Error
     def test_create_new_movie_error(self):
-        res = self.client().post('/movies', json={"title": "Men in Black2"})
+        token_Producer = os.environ['TOKEN_PRODUCER']
+        res = self.client().post('/movies', json={"title": "Men in Black2"}, headers={"Authorization":"Bearer "+token_Producer})
         data = json.loads(res.data.decode('utf-8'))
         self.assertEqual(res.status_code, 400)
         self.assertEqual(data['success'], False)
 
     #Test GET /movies
     def test_get_movies(self):
-        res = self.client().get('/movies')
+        token_Producer = os.environ['TOKEN_PRODUCER']
+        res = self.client().get('/movies', headers={"Authorization":"Bearer "+token_Producer})
         data = json.loads(res.data.decode('utf-8')) 
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     #Test GET /movies - Error
     def test_get_movies_error(self):
+        token_Producer = os.environ['TOKEN_PRODUCER']
         # Call endpoint with missing 's' at the end
-        res = self.client().get('/movie')                       
+        res = self.client().get('/movie', headers={"Authorization":"Bearer "+token_Producer})                       
         data = json.loads(res.data.decode('utf-8')) 
         self.assertEqual(res.status_code, 404)
         self.assertEqual(data['success'], False)
 
    #Test DELETE /movies
     def test_delete_movie(self):
-        setup_response = self.client().post('/movies', json={"title": "Men in Black2", "date": "2002"})
+        token_Producer = os.environ['TOKEN_PRODUCER']
+        setup_response = self.client().post('/movies', json={"title": "Men in Black2", "date": "2002"}, headers={"Authorization":"Bearer "+token_Producer})
         data_setup = json.loads(setup_response.data.decode('utf-8'))
-        res = self.client().delete('/movies/'+str(data_setup['movie_id']))
+        res = self.client().delete('/movies/'+str(data_setup['movie_id']), headers={"Authorization":"Bearer "+token_Producer})
         movie = Movies.query.filter(Movies.id==data_setup['movie_id']).one_or_none()
         data = json.loads(res.data.decode('utf-8')) 
         self.assertEqual(res.status_code, 200)
@@ -92,9 +112,10 @@ class CapstoneTestCase(unittest.TestCase):
 
    #Test DELETE /movies - Error
     def test_delete_movie_error(self):
-        setup_response = self.client().post('/movies', json={"title": "Men in Black2", "date": "2002"})
+        token_Producer = os.environ['TOKEN_PRODUCER']
+        setup_response = self.client().post('/movies', json={"title": "Men in Black2", "date": "2002"}, headers={"Authorization":"Bearer "+token_Producer})
         data_setup = json.loads(setup_response.data.decode('utf-8'))
-        res = self.client().delete('/movies/'+str(data_setup['movie_id']+2000))
+        res = self.client().delete('/movies/'+str(data_setup['movie_id']+2000), headers={"Authorization":"Bearer "+token_Producer})
         movie = Movies.query.filter(Movies.id==data_setup['movie_id']).one_or_none()
         data = json.loads(res.data.decode('utf-8')) 
         self.assertEqual(res.status_code, 404)
@@ -103,9 +124,10 @@ class CapstoneTestCase(unittest.TestCase):
 
    #Test PATCH /movies
     def test_patch_movie(self):
-        setup_response = self.client().post('/movies', json={"title": "Men in Black2", "date": "2002"})
+        token_Producer = os.environ['TOKEN_PRODUCER']
+        setup_response = self.client().post('/movies', json={"title": "Men in Black2", "date": "2002"}, headers={"Authorization":"Bearer "+token_Producer})
         data_setup = json.loads(setup_response.data.decode('utf-8'))
-        res = self.client().patch('/movies/'+str(data_setup['movie_id']), json={"title": "Fast and Furious", "date": "2002"} )
+        res = self.client().patch('/movies/'+str(data_setup['movie_id']), json={"title": "Fast and Furious", "date": "2002"}, headers={"Authorization":"Bearer "+token_Producer})
         movie = Movies.query.filter(Movies.id==data_setup['movie_id']).one_or_none()
         data = json.loads(res.data.decode('utf-8')) 
         self.assertEqual(res.status_code, 200)
@@ -114,9 +136,10 @@ class CapstoneTestCase(unittest.TestCase):
 
    #Test PATCH /movies - Error
     def test_patch_movie_error(self):
-        setup_response = self.client().post('/movies', json={"title": "Men in Black2", "date": "2002"})
+        token_Producer = os.environ['TOKEN_PRODUCER']
+        setup_response = self.client().post('/movies', json={"title": "Men in Black2", "date": "2002"}, headers={"Authorization":"Bearer "+token_Producer})
         data_setup = json.loads(setup_response.data.decode('utf-8'))
-        res = self.client().patch('/movies/'+str(data_setup['movie_id']+2000))
+        res = self.client().patch('/movies/'+str(data_setup['movie_id']+2000), headers={"Authorization":"Bearer "+token_Producer})
         movie = Movies.query.filter(Movies.id==data_setup['movie_id']).one_or_none()
         data = json.loads(res.data.decode('utf-8')) 
         self.assertEqual(res.status_code, 422)
@@ -127,38 +150,43 @@ class CapstoneTestCase(unittest.TestCase):
 
     #Test POST /actors =======================================
     def test_create_new_actor(self):
-        res = self.client().post('/actors', json={"name": "Meg Ryan", "gender": "female", "age": "32"})
+        token_Producer = os.environ['TOKEN_PRODUCER']
+        res = self.client().post('/actors', json={"name": "Meg Ryan", "gender": "female", "age": "32"}, headers={"Authorization":"Bearer "+token_Producer})
         data = json.loads(res.data.decode('utf-8'))
         self.assertEqual(res.status_code, 201)
         self.assertEqual(data['success'], True)
 
     #Test POST /actors - Error
     def test_create_new_actor_error(self):
-        res = self.client().post('/actors', json={"name": "Meg Ryan", "gender": "female"})
+        token_Producer = os.environ['TOKEN_PRODUCER']
+        res = self.client().post('/actors', json={"name": "Meg Ryan", "gender": "female"}, headers={"Authorization":"Bearer "+token_Producer})
         data = json.loads(res.data.decode('utf-8'))
         self.assertEqual(res.status_code, 400)
         self.assertEqual(data['success'], False)
 
     #Test GET /actors
     def test_get_actors(self):
-        res = self.client().get('/actors')
+        token_Producer = os.environ['TOKEN_PRODUCER']
+        res = self.client().get('/actors', headers={"Authorization":"Bearer "+token_Producer})
         data = json.loads(res.data.decode('utf-8'))
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     #Test GET /actors - Error
     def test_get_actors_error(self):
+        token_Producer = os.environ['TOKEN_PRODUCER']
         # Call endpoint with missing 's' at the end
-        res = self.client().get('/actor')                       
+        res = self.client().get('/actor', headers={"Authorization":"Bearer "+token_Producer})                       
         data = json.loads(res.data.decode('utf-8'))
         self.assertEqual(res.status_code, 404)
         self.assertEqual(data['success'], False)
 
    #Test DELETE /actors
     def test_delete_actor(self):
-        setup_response = self.client().post('/actors', json={"name": "Meg Ryan", "gender": "female", "age": "32"})
+        token_Producer = os.environ['TOKEN_PRODUCER']
+        setup_response = self.client().post('/actors', json={"name": "Meg Ryan", "gender": "female", "age": "32"}, headers={"Authorization":"Bearer "+token_Producer})
         data_setup = json.loads(setup_response.data.decode('utf-8'))
-        res = self.client().delete('/actors/'+str(data_setup['actor_id']))
+        res = self.client().delete('/actors/'+str(data_setup['actor_id']), headers={"Authorization":"Bearer "+token_Producer})
         actor = Actors.query.filter(Actors.id==data_setup['actor_id']).one_or_none()
         data = json.loads(res.data.decode('utf-8')) 
         self.assertEqual(res.status_code, 200)
@@ -167,9 +195,10 @@ class CapstoneTestCase(unittest.TestCase):
 
    #Test DELETE /actors - Error
     def test_delete_actor_error(self):
-        setup_response = self.client().post('/actors', json={"name": "Meg Ryan", "gender": "female", "age": "32"})
+        token_Producer = os.environ['TOKEN_PRODUCER']
+        setup_response = self.client().post('/actors', json={"name": "Meg Ryan", "gender": "female", "age": "32"}, headers={"Authorization":"Bearer "+token_Producer})
         data_setup = json.loads(setup_response.data.decode('utf-8'))
-        res = self.client().delete('/actors/'+str(data_setup['actor_id']+2000))
+        res = self.client().delete('/actors/'+str(data_setup['actor_id']+2000), headers={"Authorization":"Bearer "+token_Producer})
         actor = Actors.query.filter(Actors.id==data_setup['actor_id']).one_or_none()
         data = json.loads(res.data.decode('utf-8')) 
         self.assertEqual(res.status_code, 404)
@@ -179,9 +208,10 @@ class CapstoneTestCase(unittest.TestCase):
 
    #Test PATCH /actors
     def test_patch_actor(self):
-        setup_response = self.client().post('/actors', json={"name": "Meg Ryan", "gender": "female", "age": "32"})
+        token_Producer = os.environ['TOKEN_PRODUCER']
+        setup_response = self.client().post('/actors', json={"name": "Meg Ryan", "gender": "female", "age": "32"}, headers={"Authorization":"Bearer "+token_Producer})
         data_setup = json.loads(setup_response.data.decode('utf-8'))
-        res = self.client().patch('/actors/'+str(data_setup['actor_id']), json={"name": "Bud Spencer", "gender": "male", "age": "57"})
+        res = self.client().patch('/actors/'+str(data_setup['actor_id']), json={"name": "Bud Spencer", "gender": "male", "age": "57"}, headers={"Authorization":"Bearer "+token_Producer})
         actor = Actors.query.filter(Actors.id==data_setup['actor_id']).one_or_none()
         data = json.loads(res.data.decode('utf-8')) 
         self.assertEqual(res.status_code, 200)
@@ -190,136 +220,14 @@ class CapstoneTestCase(unittest.TestCase):
 
    #Test PATCH /actors - Error
     def test_patch_actor_error(self):
-        setup_response = self.client().post('/actors', json={"name": "Meg Ryan", "gender": "female", "age": "32"})
+        token_Producer = os.environ['TOKEN_PRODUCER']
+        setup_response = self.client().post('/actors', json={"name": "Meg Ryan", "gender": "female", "age": "32"}, headers={"Authorization":"Bearer "+token_Producer})
         data_setup = json.loads(setup_response.data.decode('utf-8'))
-        res = self.client().patch('/actors/'+str(data_setup['actor_id']+2000))
+        res = self.client().patch('/actors/'+str(data_setup['actor_id']+2000), headers={"Authorization":"Bearer "+token_Producer})
         actor = Actors.query.filter(Actors.id==data_setup['actor_id']).one_or_none()
         data = json.loads(res.data.decode('utf-8')) 
         self.assertEqual(res.status_code, 422)
         self.assertEqual(data['success'], False)
-
-
-
-'''  
-    #Test GET /actors
-    def test_get_paginated_questions(self):
-        res=self.client().get('/questions')
-        data = json.loads(res.data.decode('utf-8'))  
-        
-        self.assertEqual(res.status_code, 200)
-        self.assertEqual(data['success'], True)
-        self.assertTrue(data['total_questions'])
-        self.assertTrue(len(data['questions']))
-        self.assertTrue(len(data['categories']))
-
-    #Test GET /questions - Error 
-    def test_404_sent_requesting_beyond_valid_page(self):
-        res=self.client().get('/questions?page=1000', )
-        data = json.loads(res.data.decode('utf-8'))
-
-        self.assertEqual(res.status_code, 404)
-        self.assertEqual(data['success'], False)
-        self.assertEqual(data['message'], 'Not found')
-
-
-
-
-    #Test DELETE  /questions/<int:question_id>
-    def test_delete_question_by_id(self):
-        res = self.client().delete('/questions/5')
-        data = json.loads(res.data.decode('utf-8'))
-
-        question = Question.query.filter(Question.id==5).one_or_none()
-
-        self.assertEqual(res.status_code, 200)
-        self.assertEqual(data['success'], True)
-        self.assertEqual(data['deleted_question'], 5)
-        self.assertEqual(question, None)
-
-    #Test DELETE  /questions/<int:question_id>  - Error
-    def test_422_if_question_does_not_exist(self):
-        res = self.client().delete('/questions/1000')
-        data = json.loads(res.data.decode('utf-8'))
-
-        self.assertEqual(res.status_code, 422)
-        self.assertEqual(data['success'], False)
-
-
-
-    #Test POST endpoint to create new questions
-    def test_create_new_question(self):
-        res = self.client().post('/questions', json={"question":"How high is the Eiffel tower?", "answer":"300 meter", "difficulty": 4, "category":"1"})
-        data = json.loads(res.data.decode('utf-8'))
-
-        self.assertEqual(res.status_code, 200)
-
-    #Test POST endpoint to create new questions - Error 
-    def test_422_create_new_question_with_missing_json_data(self):
-        res = self.client().post('/questions', json={"answer":"300 meter", "difficulty": 4, "category":"1"})
-        data = json.loads(res.data.decode('utf-8'))
-
-        self.assertEqual(res.status_code, 422)
-        self.assertEqual(data['success'], False)
-
-    
-    #Test POST to get questions based on a search term.
-    def test_search_Term(self):
-        res = self.client().post('/questions', json={"searchTerm":"paint"})
-        data = json.loads(res.data.decode('utf-8'))
-
-        self.assertEqual(res.status_code, 200)
-        self.assertEqual(data['success'], True)
-        self.assertTrue(data['total_questions'])
-        self.assertTrue(len(data['questions']))
-
-    #Test POST to get questions based on a search term. - Error
-    def test_422_if_search_Term_does_not_match(self):
-        res = self.client().post('/questions', json={"searchTerm":"ü"})
-        data = json.loads(res.data.decode('utf-8'))
-
-        self.assertEqual(res.status_code, 422)
-        self.assertEqual(data['success'], False)
-
-
-    #Test GET endpoint to get questions based on category.
-    def test_get_questions_based_on_categories(self):
-        res = self.client().get('/categories/4/questions')
-        data = json.loads(res.data.decode('utf-8'))
-      
-        self.assertEqual(res.status_code, 200)
-        self.assertEqual(data['success'], True)
-        self.assertTrue(len(data['questions']))
-
-    #Test GET endpoint to get questions based on category.  - Error
-    def test_422_get_questions_based_on_categories_category_not_available(self):
-        res = self.client().get('/categories/1000/questions')
-        data = json.loads(res.data.decode('utf-8'))
-      
-        self.assertEqual(res.status_code, 422)
-        self.assertEqual(data['success'], False)
-
-    
-
-    #Test POST to start quizz.
-    def test_start_quizz(self):
-        res = self.client().post('/quizzes', json={"previous_questions":[20],"quiz_category":{"type":"Science","id":"1"}})
-        data = json.loads(res.data.decode('utf-8'))
-
-        self.assertEqual(res.status_code, 200)
-        self.assertEqual(data['success'], True)
-        self.assertTrue(len(data['question']))
-
-
-    #Test POST to start quizz. - Error
-    def test_422_start_quizz_no_category_in_json(self):
-        res = self.client().post('/quizzes', json={"previous_questions":[20]})
-        data = json.loads(res.data.decode('utf-8'))
-
-        self.assertEqual(res.status_code, 422)
-        self.assertEqual(data['success'], False)
-
-'''
-
 
 # Make the tests conveniently executable
 if __name__ == "__main__":
